+++ examples/03-dev/demo/project/src/test-5/index/index.js
@@ -1,12 +1,16 @@
 
 require('bootstrap/dist/css/bootstrap.css');
+// require('../../test/index/css/index.less');
+// require('../../test/index/css/index.css');
+// require('../../test/index/css/index2.css');
+require('../../less/common.less');
 require('bootstrap');
 
 require('base')();
 require('common')();
 
-// require('../../test/index/js/index2')();
-// require('../../test/index/js/index3.jsx');
+require('../../test/index/js/index2')();
+require('../../test/index/js/index3.jsx');
 
 console.log('test-5/index');
 // require('./js/b')();