+++ project_config/webpack/loaders/less_loader.js
@@ -2,11 +2,12 @@
  * make less loader
  *
  * @param useCssModules
- * @param includeNodeModules
- * @param excludeNodeModules
+ * @param cssModulesExclude
+ * @param include
+ * @param exclude
  * @returns {{test: RegExp, use: *[]}}
  */
-module.exports = (useCssModules = !1, includeNodeModules = !1, excludeNodeModules = !1) => {
+module.exports = (useCssModules = !1, cssModulesExclude = [], include = !1, exclude = !1) => {
     let loader = {
         test: /\.less$/,
         use: [
@@ -28,8 +29,8 @@ module.exports = (useCssModules = !1, includeNodeModules = !1, excludeNodeModule
         ]
     };
 
-    includeNodeModules && (loader.include = /node_modules/);
-    excludeNodeModules && (loader.exclude = /node_modules/);
+    include && (loader.include = cssModulesExclude);
+    exclude && (loader.exclude = cssModulesExclude);
 
     return loader;
 };